refactor: Updated integration_test.rs to use the new library function
